+++ CHANGELOG.md
@@ -1,16 +1,24 @@
 ## 0.8.1 (unreleased)
 
+New safe operations
+
 - Setting `NOT NULL` on an existing column is safe in MySQL 5.6+ and MariaDB 10.0+ with strict mode enabled
-- Added `target_sql_mode` option
-- Added warning for unsupported database versions
-- Added warning for `change_column_null` with default value with `safe_by_default` option
 - Changing between `text` and `citext` is safe in Postgres when not indexed
 - Changing a `varchar` column to a `citext` column is safe in Postgres when not indexed
 - Changing a `citext` column to a `varchar` column with no length limit is safe in Postgres when not indexed
 - Changing a `cidr` column to an `inet` column is safe in Postgres
 - Increasing the precision of an `interval` or `time` column is safe in Postgres
+
+New unsafe operations
+
 - Decreasing the precision of a `timestamp` or `timestamptz` column is not safe in Postgres
 - Passing a default value to `change_column_null` is not safe in Postgres
+
+Other
+
+- Added warning for unsupported database versions
+- Added warning for `change_column_null` with default value with `safe_by_default` option
+- Added `target_sql_mode` option
 
 ## 0.8.0 (2022-02-09)