+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.8.1 (unreleased)
+## 0.9.0 (unreleased)
 
 New safe operations with MySQL and MariaDB